Deal better with types that implement Positional
A type that implemented Positional but has accessors (so presumably is
amenable to a pairwise constructor,) wouldn't have worked before.

This may not work for a Positional which has a constructor which takes
a list of things but by that point I'd expect the implementor to provide
their own BUILD or custom constructor.